requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
mysql-connector-python==8.2.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pandas==2.1.3
openpyxl==3.1.2
python-magic==0.4.27
aiofiles==23.2.1
bcrypt==4.1.2
python-dateutil==2.8.2